+++ src/backend/commandLine/CommandLine.cpp
@@ -45,6 +45,7 @@ CommandLine::~CommandLine()
     delete mSaveOption;
     delete mSaveToOption;
     delete mVersionOption;
+    delete mUploadOption;
 }
 
 void CommandLine::addImageGrabberOptions(const QList<CaptureModes> &captureModes)
@@ -79,6 +80,7 @@ void CommandLine::addDefaultOptions()
     mEditOption = addParameterOption(QLatin1String("e"), QLatin1String("edit"), QLatin1String("Edit existing image in ksnip."), QLatin1String("image"));
     mSaveOption = addOption(QLatin1String("s"), QLatin1String("save"), QLatin1String("Save screenshot to default location without opening in editor."));
     mSaveToOption = addParameterOption(QLatin1String("p"),QLatin1String("saveto"),QLatin1String("Save screenshot to provided path without opening in editor."), QLatin1String("path"));
+	mUploadOption = addOption(QLatin1String("o"), QLatin1String("upload"), QLatin1String("Upload screenshot via default uploader without opening in editor."));
 }
 
 void CommandLine::addVersionOptions()
@@ -185,6 +187,11 @@ QString CommandLine::saveToPath() const
 bool CommandLine::isCaptureModeSet() const
 {
     return isRectAreaSet() || isLastRectAreaSet() || isFullScreenSet() || isCurrentScreenSet() || isActiveWindowSet() || isWindowsUnderCursorSet();
+}
+
+bool CommandLine::isUploadSet() const
+{
+	return mUploadOption != nullptr && isSet(*mUploadOption);
 }
 
 CaptureModes CommandLine::captureMode() const